Updated ReadMe and Removed Useless Files
Updated the readme with info on most of the files and folders needed for the server. Also removed the MC Forge installer files which are no longer needed, and streamlined the name of the Forge server launcher. This then meant the server_start.command needed updating to reflect the change in the Forge launcher name.