Added One More Line to ReadMe
Added one more line to the first section of the readme to explain why this repo exists. There wouldn't be much point to just storing a Minecraft server in a repo unless the machine that will be running the server is remote. This could be managed with FTP, but this way I can have a copy of the server folder on my own machine and modify as I see fit then simply push that to the remote host.