Initialized Server
I actually ran the Forge jar file to start the server and generate the remaining server files. This also generated a new map folder "Biome Bundle" which contains the world with modified terrain generated by the OTG preset. Also added the ".mixin.out" folder to the ignored files after it came back and I did a little research on it. I will let the remote host generate its own and store it locally as all it has to do with is debugging the Sponge API.