Finished Detailed Structure
Finished up the detailed file structure section. There are a few items that I do not know, and have not found descriptions of in the various documentations. There are also a few that it would just be silly to actually dive all the way into, IE java libraries, so I simply left them with a higher level description.